refactor: PromotionService 네이밍 및 패키지 변경
